refactor: 폴더 이동
- database-connect를 middleware에서 lib으로 이동
- 기능에 맞도록 폴더 이동
